+++ PecuniaFinanceSystem1/src/main/java/com/capgemini/pecunia/service/LoanRequestServiceImp.java
@@ -12,7 +12,7 @@ public class LoanRequestServiceImp implements LoanRequestService {
 	LoanRequestDao dao;
 	@Override
 	public String loanRequest(LoanRequests loanreq) {
-		long s1 = loanreq.getAccountId();
+		int s1 = loanreq.getAccountId();
 		Optional<Account> details = dao.findBank(s1);
 		if (details.isPresent()) {
 			dao.save(loanreq);
